Fix tensorflow import
